+++ src/Response/InternalTagResponse.php
@@ -1,0 +1,36 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Storyblok\ManagementApi\Response;
+
+use Storyblok\ManagementApi\Data\InternalTag;
+use Storyblok\ManagementApi\Exceptions\StoryblokFormatException;
+use Symfony\Contracts\HttpClient\Exception\ClientExceptionInterface;
+use Symfony\Contracts\HttpClient\Exception\DecodingExceptionInterface;
+use Symfony\Contracts\HttpClient\Exception\RedirectionExceptionInterface;
+use Symfony\Contracts\HttpClient\Exception\ServerExceptionInterface;
+use Symfony\Contracts\HttpClient\Exception\TransportExceptionInterface;
+
+class InternalTagResponse extends StoryblokResponse implements StoryblokResponseInterface
+{
+    /**
+     * @throws TransportExceptionInterface
+     * @throws ServerExceptionInterface
+     * @throws RedirectionExceptionInterface
+     * @throws DecodingExceptionInterface
+     * @throws StoryblokFormatException
+     * @throws ClientExceptionInterface
+     */
+    #[\Override]
+    public function data(): InternalTag
+    {
+        $key = "internal_tag";
+        $array = $this->toArray();
+        if (array_key_exists($key, $array)) {
+            return InternalTag::make($array[$key]);
+        }
+
+        throw new StoryblokFormatException(sprintf("Expected '%s' in the response.", $key));
+    }
+}